change the name of the window when windows are swapped
This allows for the name of the gui (displayed in the title bar) to
be displayed when the window is changed. In addition, a small
addition to the window structure is made, to mark tag windows as such,
allowing them to be ignored for tasks such as getting the filename.